merging rev no 11821 from 6.1/trunk MEB Fixes: output handling, remote temorary files, config exception handling. - Fixed output handling on command execution: was not being passed back on sudo command failure. - Updated remote temporary file logic to not write to the user's home but to a random generated file on /tmp. - Fixed exception handling for file save operation on the meb configuration settings.
Fixes Bug #18676962 - MEB FAILS WITH AN UNHANDLED EXCEPTION MESSAGE BOX
Problem: A wrong exception handler was being used when saving the configuration file which bypassed the error.
Fix: Being the UI, the exception handled was replaced by a generic which won't bypass any error. Logging of the exception was also added.

[TESTS]
- Ensured MEB is able to create the configuration file even the user's home is on an nfs mount.