- Fabric: - Adding fabric to the windows solution               - Updating the PrepareOutput scripts to include the new fabric module and the fabric utilities code
- Fabric Module : - Adding patch to add getheader method to addurlinfo class, which is missing in Python 2.7 and caused crash.
                           - Updating authentication to remove unneeded connection information for the backing store.
                           - Creating the managed server connections using the user provided in the options tab of the connection settings

- Connection Editor: - Removing unneded fields for the  backing store from the connection settings for the fabric connection type.
                                 - Added Options tab for the fabric connection type, includes a field for the mysql user for the managed server connections.


[TESTS]
Ensured the connector editor displays the options tab only for fabric connection type and that it contains the fild for the mysql user for the managed servers.
Ensured the TestConnection works properly.
Ensured the Connections for the managed servers are created successfully when a connection to a fabric node is selected.
Ensured the connections for the managed servers are created using the user set on the options tab for the fabric connection.