Fabric: Misc updates/fixes
- Updated some captions to make them more clear or descriptive.
- Fixed caption of the Local Socket/Pipe conection type.
- Fixed bug creating configuration for Fabric module operations.
- Fixed handling when user cancelled password dialog.
- Added driver_changed callback to the New Connection Wizard to properly show/hide the Server Configuration button.
  based on the selected connection type: On Fabric connection it will hidden.
- Displayed the Wait Panel while creating the connections to the managed MySQL Servers.
- Displayed status messages while creating the connections to the managed MySQL Servers.
- Added validation to not display the Group Filter field on the popup info box if no filter is defined.

[TESTS]
Updated captions were used on the UI.
Unhandled exception when testing connection but cancelling the password input was no longer shown.
Server Configuration Button is hidden if the fabric configuration type is selected, and shown otherwise.
When the managed connections are being created, the wait dialog is shown and the status bar reflects what´s going on.
When no group filter is set for the connection, the field is not shown on the popup info box.