MEB: Fixing integration of the MEB Helper
Issue Description
Found a situation of mutual exclusion on the location of the helper file. It was being used as a module for the version validation and as a resource when deployed on the backups home. It happens that if located at ${WB_PACKAGE_SHARED_DIR}/libraries it is available as a module but not as a resource to be deployed, if located on ${WB_PACKAGE_SHARED_DIR} it is available as resource to be deployed but not as a module for version validation.
Renamed the MEBVersion class on the helpers to MEBHelperVersion to avoid confusion.

Change:
File was finally located at ${WB_PACKAGE_SHARED_DIR} to make it a valid resource and version validation was changed to be done independent of the helper file, following the same schema as the windows helper.

[TESTS]
Ensured the helper is now deployable and that the version validation works.
Tested fresh deploymentand update deployment of the MEB Helper.