PYTHONCONTEXT: Adding way to set additional paths for the PYTHONPAT
Have added reading of a new environment variable WB_PYTHONPATH, so we can configure there additional path to python libraries so they can be used inside WB.

This is specially useful for out own testing purposes.

[TESTS]
Pointed the env variable to the system python lib folders and we were able to use unittest and connector/python modules which are not part of the embedded python.
Inside the scripting shell imported a module that is in both the embedded python and the system install and it imported the embedded one.