Win: solved trouble with recreating the auto completion cache
Until now if there was a change in the structure of the auto completion cache it was recreated by closing the cache, removing the file and reopening it. That often failed on Windows. However, we refresh all cache content on creation of the auto completion class anyway, so I changed the code that we always try to remove the cache file before even opening it.

Additionally, I added a tryRemove() file function that doesn't throw an exception if removing fails (returns false instead).

review by: lolek [rb: 8512]