MW-1041 VALID STORED PROCEDURE CODE LABELLED _SYNTAX_ERROR IN ROUTINE GROUP
Handling of name scanning of routine names was wrong if the code contained a syntax error, leading to a wrong object name. Now handling is done using a scanner as that doesn't need a valid AST (which we don't have in error case).

En passant I changed the return type from createScanner in the parser module to use a shared_ptr instead of a raw one.

Review by: lolek (rb: #9219)