Fix for a parser problem with the new generated columns syntax
Because of the addition of the generated column syntax a new problem appeared, rendering all standard datatypes as syntax errors if their first attribute was a collation clause (as this clause is also needed in the gc support).

Additionally, I updated the code editor xml to support all new keywords for 5.7 and fixed comment parsing in the parser module (the comment string is a sub tree, not a single token). This also broke table editing with comments in columns.

Reviewed by: lolek [rb #10539]