Mac: rework of canvas drawing
Starting with OSX 10.11 the handling of core graphics contexts is more strict. There is no longer a valid context outside of a paint cycle. WB's canvas setup was designed however to use a fixed context set upfront, which no longer worked. We have a very similar situation on Windows, so I changed the implementation to very closely follow that on Win.

Review by: lolek (rb: #10701)