TUT: fixed crash in sql_editor_be_autocomplete_tests
Added settings to make debugging wbtests possible in XCode. They PATH adjustment is machine dependend as it depends on the actual server used, but that is easily adjustable.
Also removed the code_editor.xml file from repo entirely, as it is copied from the source folder for testing.

Review by: Miguel