MW-1291 WB DOESN'T RECOGNIZE VALID ALTER USER SYNTAX
Fixed a number of issues in the ANTLR grammar that lead to wrong recognition. Converted a number of gated semantic predicates to ungated ones as this makes a difference in recognition as well (and also error reporting).

While I was working in the grammar test app I also applied recommended changes from XCode.

Review by: Miguel