SQL IDE: added support for version + sql mode hints in error messages
With this patch an additional line is added to parser error information that gives a hint why a particular error occured, if it depends on a specific server version or an SQL mode being (not) active. This also requires that non-gated syntactic predicates are used in the ANTLR grammar (because only then we get the necessary predicate error).

Reviewed by: Miguel (rb: #11026)