MW-1340 [MODEL] EDITING A ROUTINE IN ROUTINE GROUP EDITOR DUPLICATES ROUTINE OBJECT
When editing a routine group the parser didn't find an existing routine and duplicated it, caused by not-removed quotes. This is a special case for routines because it uses a dedicated function to get name + type of a routine without using the parser (as it is also used for errorneous queries).

Additionally, a few grammar changes that had been forgotten were migrated from the main grammar to the simple grammar.

Review by: Lolek (rb #11189)