MW-1339 [MODEL] DESCRIPTION PANE DOESNT WORK FOR MULTIPLE SELECTED OBJECTS ON OVERIVEW
Multiselection is actually not supported on the model overview page (at least on OSX), but some mouse handling code allowed it, which lead to a discrepancy. This handling has been updated to consider the multi selection mode, but must be enhanced later to fully allow multiselection.

Also updated a few other places (copyright, owner property) and removed 2 bindings to a no longer existing var/property.