Some cleanup in recordset
Removed the data_edited() call in RecordSet::reset() as this is frequently called from a background thread and the signal is used to update the UI. In fact there is no need anyway, since the reset call is always only the preparation for a following fill up with records.

Also removed an unneeded lock for the reset call in the SQL IDE. That call no longer needs the aux connection.

Review by: lolek (rb #11205)