Fixed GRANT rule in ANTLR grammar
The grant rule (more precisely the grant_user rule) incorrectly allowed the IDENTIFIED BY part only for servers >= 5.7.6, while in fact this is available at least since 5.1, if not earlier. This has been corrected.
Additionally, the parseStatemntDetails function in the parser module returns now an error message if it was asked for an unsupported query type.

Review by: lolek, milosz (rb #11266)