MW-1409 ERROR PARSING VALID STORED PROCEDURE
This patch implements the feature that all identifiers (including keywords) can appear unquoted after a dot in a qualified identifier. In the server this is done using handwritten lexer code. Here we need to implement that in the grammar.

This patch contains a number of accompanying changes:
- A no longer used function in the parser test app has been removed.
- The predefined token list now contains all keywords, not only those from the keyword + keyword_sp rules. This is necessary to have an efficient check for any keyword.
- Moved the dumpTree instance function from mysql-parser.cpp to mysql-parser-common.cpp as static function to make it more universally usable.
- New syntax tests have been added that check various qualified and non-qualified identifiers. A number of very similar tests have been removed.

Review by: lolek (rb #11558)