Mac: another attempt at making TUT tests working again in this branch.
Due to the conversion to ARC and accompanying changes it was necessary to reimplement the way wbtests stored references of stub mforms objects (which stand for platform objects, but we don't have such in wbtests).
This also required to put platform specific code for stub::ObjectWrapper into an own file.

This patch also fixes a wrong merge (pre-ARC code left in the branch).