MW-1815 Workbench freezes when adding inserts to a table in an EER diagram
The apply code ran on a background task but waited for it, blocking so the main thread. The background task tried to update the UI, causing a deadlock. Waiting (blocking) for a background task means we actually don't need a background task. So, the apply code now runs in main thread.

Note: there is currently no feedback that the data has been applied. That should be handled by caller which applied the changes actually (which requires some extra rework).