Change to grt dispatcher callback waiting
This patch exchanges the old GCond implementation by our Semaphore to avoid trouble when the free signal was sent before the callback started to wait. This is a race condition we did not consider so far. With the semaphore this changes now as it keeps its "signaled" state even if set before the wait started (in that case the grt task can immediately return, since everything is done already).