Bug #27314770 DEFINE A COLUMN THAT INCLUDES FSP CAUSE "ALTER TABLE PANEL" RENDER INCOMPLETE
This bug was already fixed with the parser rework, however the table editor didn't show timestamp column types with length in the columns list (also not for default values). This patch fixes that.

Additionally I applied proper formatting to the table editor .mm file and let XCode apply it's new defaults to the table editor xib.