Bug #25510537 UPDATE/DELETE GENERATED QUERY IS INCORRECT WHEN UPDATING A ROW WITH A NULL VALUE
The generated SQL was wrong when updating a field that was originally NULL. The predicate code for the update command unconditionally used comparison. This is has been changed to use IS NULL for null values instead.

Additionally, queries are now clearer (added some spacing and parentheses).

Finally, there was a warning from XCode that UI code was accessed from a background thread while running the apply wizard. This has been fixed by calling that code from the main thread.