Workaround for bad design
This patch works around the design of the record set class, where GUI code is called from background threads. A part of that has been moved to an idle task, to be run on the main thread. This causes trouble sometimes however, namely when the recordset is freed before the callback was triggered (and there's no way of removing a once scheduled callback, another design problem). So for now we don't trigger that callback at all, if we are unit-testing.