Bug #27534644 [MASTER][SQLED] UNABLE TO CREATE A SCHEMA
A regression had been introduced on macOS after a refactoring which made the schema editor not allowing to change the name of a new schema in the SQL IDE. This has been fixed.

Along with that the schema editor has been overhauled. It was using only a very small part of the available space, with very small fonts and bad layout. This has been improved and the wording updated as well.