Follow up fixes for WL#10634
The rework of the admin tabs left them in a bad state on macOS, due to some pecularities there. This patch changes some of the GUI construction code to take these pecularities into account when running on macOS. Additionally, I had to fix the tab drawing for secondary tabs on macOS as the old code did not properly account for title sizes and available space.