Fix for model sync
This patch should have been part of the parser refactoring. Previously the sqlDefinition member of a DB grt object contained only the "body" of a statement (which is a bit vague, since that can differ). Only having the body there was necessary to make object editors work (where a part of the query was generated and the rest taken from the sqlDefinition). After refactoring this is no longer needed, hence we can include the full SQL query text there.

However, this requires to fix a case in diff where previously a complete query was created out of the "body" text. We simply can take what's in sqlDefinition now instead.