Bug #27910024 	MYSQL WORKBENCH CRASHES WHENEVER TYPING CHANGE MASTER STATEMENT
This was a problem wrt code completion, which runs also grammar predicates. In this specific situation a parse context was needed, which doesn't exist during code completion. After all checking the semantics during the parse step is not a good idea either (in the comment there's already a hint), so the predicate has been removed.