Improve messages for parse errors in the SQL IDE
This patch improves the messages shown for SQL errors in editors significantly by considering the parser context where the syntax error occured and by considering the special case of certain keywords allowed as identifiers in certain circumstances, which led to errors listing these keywords, while in fact an identifier was required.

Furthermore, the error messages now contain very specific infos in certain situations like specifying a name for new tables, views, schemas etc. or when referencing existing ones. Also function calls are recognized and described as incomplete if so.