Fixed some grammar errors and a crash
A few grammar problems have been fixed with this patch:
- ODBC table reference syntax allowed recursion.
- The server version was wrong for the JSON table function in tableFactor.
- Corrected joinedTableParens and found that it is actually a duplicate of tableReferenceListParens.
- The alias for JSON_TABLE is optional.

This require a small change in the parser module (without functional change).

Additionally, a crash was fixed in code completion because a select with only derived table references isn't handled there yet and led to an empty reference stack (that crashed on access).