Bug #28222529 [ADMIN] CLOSING A SSH CONNECTION AFTER STOP A SERVICE CRASHES WB
remove lock in the sftp->open() that caused the deadlock.
this function is used only in the SSHSessionWrapper::open which already holds the lock.

Refactor handling of pollSession(), now we create ThreadedTimer once and remove it in the d-tor.
ThreadedTimer::remove return bool value depends if the task was on the list/pool or not so the caller can wait on the task if needed.