wl-12381 Focusable areas on mforms DrawBox
In order to support accessibility properly we need a way to declare areas in a draw box to be focusable (e.g. for tiles on the WB homescreen which act as buttons and are also reported as such to Accessibility). Such areas should behave as if they were normal child elements (e.g. wrt auto scrolling into view, reacting to keyboard input like <space> or <enter> etc.).