Bug #29547778 [FIREWALL] CAN'T RE-ENABLE FIREWALL AFTER UNINSTALLATION This is fixed due to proper handling of the install process
Bug #29681550 EXCEPTION IS SHOWN WHEN RESETING FIREWALL RULES
This is fixed by disabling the users firewall tab when the firewall is removed, since it's not possible to add/remove the firewall additions.

Bug #23617930 WORKBENCH REFUSES TO CONTROL THE FIREWALL ON A DEVELOPMENT SERVER
No need to check if the server is commercial or community. Checking for the  mysql_firewall_mode is enough, since it will only be available when the firewall is installed, and it only installs on the commercial server.

Fixed mode check, which in some cases was failing.
Fixed removing the footer in the admin tabs
Added missing firewall files to the VS project
Never show failing dialog (just log) when uninstalling, since it's unnecessary for the user.