Since this repo doesn't have "issues". I'll do it here: YOU CANT REPORT BUGS A SIMPLE WAY QUICKLY
When you get an error in your MySQL Workbench, it asks you to report or not a bug. So here we go.
Oracle just asks a FUCKING WHOLE PAGE FULL OF BUROCRACY TO FILL just to report a FUCKING BUG

You guys just can't do a simple way to report bugs here? WITHOUT ORACLE WONDERING TO KNOW WHERE I LIVE (WTFFFFFF) to just