ET#72670: Build against bundled Python 3.9.7 on Fedora 35
The Fedora 35 build was added after code freeze and is the only
supported platform that uses Python 3.10. Changes in Python 3.10 to
remove the old parser module makes the build fail. As a temporary
solution, Python 3.9.7 is bundled in Workbench 8.0.27.

The already existing solution for EL7 was made more flexible, to be
able to specify to "rpmbuild" the Python include directory and the
Python library to bundle.

In Fedora 35 the "rpmbuild" RPATH check is enabled as a default. This
test is disabled until the 3rd party "ogr2ogr" and "ogrinfo" builds
are corrected.