Fix buffer overflows on several places (Scintilla)
The issue is that there was the possibility to write 1 byte after the buffer, since the while(len<size) would still enter when len == size - 1. So the rational should be while(len<size-1).

To do so, const length variables were added, so that we don't have magic numbers there. That length is now passed into the functions, so they can use those vars.