Adding support to embedd the MySQL Shell Workbench (SWB)
This patch updates the packaging scripts to support embedding
the MySQL Shell Workbench (SWB) as part of the Workbench package.

The following environment variables are required to achieve it:

SWBWIXPATH: Path to the WIX configuration for the SWB as a component.
SWBPACKAGE: Path to the SWB binaries.

If the variables are not defined, the Workbench packages will
have no changes.